AUTH is completed, orders are completed, todo/ authors genre, flyway, exceptionhandlers

+++ project2-branch2/src/main/java/com/example/bookstore/services/implementations/AuthorServiceImpl.java
@@ -1,29 +1,31 @@
 package com.example.bookstore.services.implementations;
 
-import com.example.bookstore.DTOs.Author.AuthorRequestDto;
-import com.example.bookstore.DTOs.Author.AuthorResponseDto;
-import com.example.bookstore.Repos.AuthorRepo;
+import com.example.bookstore.dto.author.AuthorRequestDto;
+import com.example.bookstore.dto.author.AuthorResponseDto;
+import com.example.bookstore.repository.AuthorRepo;
+import com.example.bookstore.repository.GenreRepo;
 import com.example.bookstore.entities.Author;
+import com.example.bookstore.exceptions.ResourceNotFoundException;
 import com.example.bookstore.services.AuthorService;
 import com.example.bookstore.services.BookService;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class AuthorServiceImpl implements AuthorService {
 
     private final AuthorRepo authorRepo;
+    private final GenreRepo genreRepo;
     private final BookService bookService;
-        public AuthorServiceImpl(AuthorRepo authorRepo, BookService bookService) {
-        this.authorRepo = authorRepo;
-            this.bookService = bookService;
-        }
 
-    private static Throwable get() {
-        return new IllegalArgumentException("Author with id not found");
+    public AuthorServiceImpl(AuthorRepo authorRepo, GenreRepo genreRepo, BookService bookService) {
+        this.authorRepo = authorRepo;
+        this.genreRepo = genreRepo;
+        this.bookService = bookService;
     }
 
     @Override
@@ -33,53 +35,73 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public Optional<AuthorRequestDto> getByID(Long id) {
-        return authorRepo.findById(id).map(Author::convertAuthorToRequestDto);
+        if (authorRepo.existsById(id))
+            return authorRepo.findById(id).map(Author::convertAuthorToRequestDto);
+        else throw new ResourceNotFoundException("The author with ID %d is not found", id);
     }
+
     @Override
     public List<AuthorRequestDto> getByName(String name) {
-       return authorRepo.findByName(name).stream().map(Author::convertAuthorToRequestDto).toList();   }
+        if (authorRepo.existsByNameIsContainingIgnoreCase(name) ||
+                authorRepo.existsBySurnameIsContainingIgnoreCase(name) ||
+                authorRepo.existsByPatronymicIsContainingIgnoreCase(name))
+            return authorRepo
+                    .findByNameIsContainingIgnoreCaseOrSurnameIsContainingIgnoreCaseOrPatronymicIsContainingIgnoreCase(name, name, name)
+                    .stream()
+                    .map(Author::convertAuthorToRequestDto)
+                    .toList();
+        else throw new ResourceNotFoundException("Author is not found or doesn't exist");
+    }
 
     @Override
-    public List<AuthorResponseDto> getAuthorsByGenreName(String name) {
-        List<Author> authorList=new ArrayList<>();
-        bookService.getAuthorByGenreName1(name)
-                .forEach(book -> authorList.addAll(book.getAuthorList().stream().toList()));
-        return authorList.stream().map(Author::convertAuthorToResponseDto).toList();
-
+    public Set<AuthorResponseDto> getAuthorsByGenreName(String name) {
+        if (!genreRepo.existsByGenreName(name))
+            throw new ResourceNotFoundException("The genre %s is not found or doesn't exist", name);
+        else {
+            return authorRepo.findAllByGenre(name).stream().map(Author::convertAuthorToResponseDto).collect(Collectors.toSet());
+//            List<Author> authorList = new ArrayList<>();
+//            bookService.getAuthorByGenreName1(name)
+//                    .forEach(book -> authorList.addAll(book.getAuthorList().stream().toList()));
+//            return authorList.stream().map(Author::convertAuthorToResponseDto).collect(Collectors.toSet());
+        }
     }
+
     @Override
     public void deleteByID(Long id) {
-         authorRepo.deleteById(id);
+        if (authorRepo.existsById(id))
+            authorRepo.deleteById(id);
+        else throw new ResourceNotFoundException("The author with ID %d doesn't exist", id);
     }
 
     @Override
     public void create(AuthorResponseDto authorResponseDto) {
 
         authorRepo.save(
-               new Author(
-                       authorResponseDto.getName(),
-                       authorResponseDto.getSurname(),
-                       authorResponseDto.getPatronymic(),
-                       authorResponseDto.getDateOfBirth()
+                new Author(
+                        authorResponseDto.getName(),
+                        authorResponseDto.getSurname(),
+                        authorResponseDto.getPatronymic(),
+                        authorResponseDto.getDateOfBirth()
 
-                )
-       );
+                        )
+        );
 
     }
+
     @Override
     public void update(AuthorResponseDto authorResponseDto) throws Throwable {
-            Author existingAuthor;
-            Author author=authorResponseDto.convertAuthorRequestDtoToEntity();
-            existingAuthor = authorRepo.findById(author.getId()).orElseThrow();
-            existingAuthor.setName(author.getName());
-            existingAuthor.setSurname(author.getSurname());
-            existingAuthor.setPatronymic(author.getPatronymic());
-            existingAuthor.setDateOfBirth(author.getDateOfBirth());
-            existingAuthor.setId(author.getId());
+        Author existingAuthor;
+        Author author = authorResponseDto.convertAuthorRequestDtoToEntity();
+        existingAuthor = authorRepo.findById(author.getId()).orElseThrow(() ->
+                new ResourceNotFoundException("The author with ID %d is not found", author.getId()));
+        existingAuthor.setName(author.getName());
+        existingAuthor.setSurname(author.getSurname());
+        existingAuthor.setPatronymic(author.getPatronymic());
+        existingAuthor.setDateOfBirth(author.getDateOfBirth());
+        existingAuthor.setId(author.getId());
 
-            authorRepo.save(author);
+        authorRepo.save(author);
     }
-
 
 
 }
